test(acp): keep runtime contract in existing shard
