+++ src/shared/acp.test.ts
@@ -1,6 +1,30 @@
 import { describe, expect, it } from 'vitest'
 
 import { sanitizeAcpModelCallUsage, toAcpTurnTokenUsage } from './acp'
+import type { AcpRuntimeEvent } from './acp'
+
+const eventBase = {
+  id: 'event-1',
+  timestamp: 1,
+  level: 'info'
+} as const
+
+// @ts-expect-error artifact events require their run and artifact payload
+const artifactWithoutPayload: AcpRuntimeEvent = { ...eventBase, kind: 'artifact' }
+
+// @ts-expect-error message events require a role and text
+const messageWithoutContent: AcpRuntimeEvent = { ...eventBase, kind: 'message' }
+
+// @ts-expect-error permission events require their request identity
+const permissionWithoutRequestId: AcpRuntimeEvent = { ...eventBase, kind: 'permission' }
+
+// @ts-expect-error tool events cannot carry artifact payloads
+const toolWithArtifactPayload: AcpRuntimeEvent = {
+  ...eventBase,
+  kind: 'tool',
+  toolCallId: 'tool-1',
+  artifacts: []
+}
 
 describe('ACP turn token usage', () => {
   it('preserves cache details only when the agent reports both read and write categories', () => {
@@ -48,5 +72,16 @@ describe('ACP turn token usage', () => {
       outputTokens: 3
     })
     expect(sanitizeAcpModelCallUsage({ ...call, id: '   ' })).toBeUndefined()
+  })
+})
+
+describe('AcpRuntimeEvent contract', () => {
+  it('rejects payloads that do not match their event kind', () => {
+    expect([
+      artifactWithoutPayload,
+      messageWithoutContent,
+      permissionWithoutRequestId,
+      toolWithArtifactPayload
+    ]).toHaveLength(4)
   })
 })
